# docs/swagger.yaml
basePath: /api
definitions:
...
  domain.Apartment:
    properties:
      agreement_accepted_at:
        type: string
      amenities:
        items:
          $ref: '#/definitions/domain.PopularAmenities'
        type: array
      apartment_number:
        type: integer
      apartment_type:
        $ref: '#/definitions/domain.ApartmentType'
      apartment_type_id:
        type: integer
      booking_count:
        type: integer
      building:
        type: string
      city:
        $ref: '#/definitions/domain.City'
      city_id:
        type: integer
      condition:
        $ref: '#/definitions/domain.ApartmentCondition'
      condition_id:
        type: integer
      contract_id:
        type: integer
      created_at:
        type: string
      daily_price:
        type: integer
      description:
        type: string
      district:
        $ref: '#/definitions/domain.District'
      district_id:
        type: integer
      documents:
        items:
          $ref: '#/definitions/domain.ApartmentDocument'
        type: array
      floor:
        type: integer
      house_rules:
        items:
          $ref: '#/definitions/domain.HouseRules'
        type: array
      id:
        type: integer
      is_agreement_accepted:
        type: boolean
      is_favorite:
        type: boolean
      is_free:
        type: boolean
      kitchen_area:
        type: number
      listing_type:
        type: string
      location:
        $ref: '#/definitions/domain.ApartmentLocation'
      microdistrict:
        $ref: '#/definitions/domain.Microdistrict'
      microdistrict_id:
        type: integer
      moderator_comment:
        type: string
      owner:
        $ref: '#/definitions/domain.PropertyOwner'
      owner_id:
        type: integer
      photos:
        items:
          $ref: '#/definitions/domain.ApartmentPhoto'
        type: array
      price:
        type: integer
      rental_type_daily:
        type: boolean
      rental_type_hourly:
        type: boolean
      residential_complex:
        type: string
      room_count:
        type: integer
      service_fee_percentage:
        type: integer
      status:
        $ref: '#/definitions/domain.ApartmentStatus'
      street:
        type: string
      total_area:
        type: number
      total_floors:
        type: integer
      updated_at:
        type: string
      view_count:
        type: integer
    type: object
...
  domain.Booking:
    properties:
      apartment:
        $ref: '#/definitions/domain.Apartment'
      apartment_id:
        type: integer
      booking_number:
        type: string
      can_extend:
        type: boolean
      cancellation_reason:
        type: string
      cleaning_duration:
        type: integer
      contract_id:
        type: integer
      created_at:
        type: string
      door_status:
        $ref: '#/definitions/domain.DoorStatus'
      duration:
        type: integer
      end_date:
        type: string
      extension_duration:
        type: integer
      extension_end_date:
        type: string
      extension_price:
        type: integer
      extension_requested:
        type: boolean
      final_price:
        type: integer
      id:
        type: integer
      is_contract_accepted:
        type: boolean
      last_door_action:
        type: string
      owner_comment:
        type: string
      payment_id:
        type: integer
      renter:
        $ref: '#/definitions/domain.Renter'
      renter_id:
        type: integer
      service_fee:
        type: integer
      start_date:
        type: string
      status:
        $ref: '#/definitions/domain.BookingStatus'
      total_price:
        type: integer
      updated_at:
        type: string
    type: object
...
  domain.ChatMessage:
    properties:
      chat_room:
        $ref: '#/definitions/domain.ChatRoom'
      chat_room_id:
        type: integer
      content:
        type: string
      created_at:
        type: string
      edited_at:
        type: string
      file_name:
        type: string
      file_size:
        type: integer
      file_url:
        type: string
      id:
        type: integer
      read_at:
        type: string
      reply_to:
        $ref: '#/definitions/domain.ChatMessage'
      reply_to_id:
        type: integer
      sender:
        $ref: '#/definitions/domain.User'
      sender_id:
        type: integer
      status:
        $ref: '#/definitions/domain.MessageStatus'
      type:
        $ref: '#/definitions/domain.MessageType'
      updated_at:
        type: string
    type: object
  domain.ChatRoom:
    properties:
      apartment:
        $ref: '#/definitions/domain.Apartment'
      apartment_id:
        type: integer
      booking:
        $ref: '#/definitions/domain.Booking'
      booking_id:
        type: integer
      closed_at:
        type: string
      concierge:
        $ref: '#/definitions/domain.Concierge'
      concierge_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      last_message_at:
        type: string
      opened_at:
        type: string
      renter:
        $ref: '#/definitions/domain.Renter'
      renter_id:
        type: integer
      status:
        $ref: '#/definitions/domain.ChatRoomStatus'
      updated_at:
        type: string
    type: object
  domain.ChatRoomStatus:
    enum:
    - pending
    - active
    - closed
    - archived
    type: string
    x-enum-varnames:
    - ChatRoomStatusPending
    - ChatRoomStatusActive
    - ChatRoomStatusClosed
    - ChatRoomStatusArchived
...
  domain.Concierge:
    properties:
      apartments:
        items:
          $ref: '#/definitions/domain.Apartment'
        type: array
      created_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      schedule:
        $ref: '#/definitions/domain.ConciergeSchedule'
      updated_at:
        type: string
      user:
        $ref: '#/definitions/domain.User'
      user_id:
        type: integer
    type: object
...
  domain.CreateChatRoomRequest:
    properties:
      booking_id:
        type: integer
    required:
    - booking_id
    type: object
...
  domain.DoorStatus:
    enum:
    - closed
    - open
    type: string
    x-enum-varnames:
    - DoorStatusClosed
    - DoorStatusOpen
  domain.ErrorResponse:
    properties:
      error:
        type: string
      success:
        type: boolean
    type: object
...
  domain.MarkMessagesReadRequest:
    properties:
      message_ids:
        items:
          type: integer
        type: array
    required:
    - message_ids
    type: object
  domain.MessageStatus:
    enum:
    - sent
    - delivered
    - read
    type: string
    x-enum-varnames:
    - MessageStatusSent
    - MessageStatusDelivered
    - MessageStatusRead
  domain.MessageType:
    enum:
    - text
    - image
    - file
    - system
    - welcome
    - status
    type: string
    x-enum-varnames:
    - MessageTypeText
    - MessageTypeImage
    - MessageTypeFile
    - MessageTypeSystem
    - MessageTypeWelcome
    - MessageTypeStatus
...
  domain.Renter:
    properties:
      created_at:
        type: string
      document_type:
        $ref: '#/definitions/domain.DocumentType'
      document_url:
        additionalProperties:
          type: string
        type: object
      id:
        type: integer
      photo_with_doc_url:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/domain.User'
      user_id:
        type: integer
      verification_status:
        $ref: '#/definitions/domain.VerificationStatus'
    type: object
  domain.SendMessageRequest:
    properties:
      content:
        type: string
      reply_to_id:
        type: integer
      type:
        $ref: '#/definitions/domain.MessageType'
    required:
    - content
    - type
    type: object
...
  domain.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  domain.UpdateMessageRequest:
    properties:
      content:
        type: string
    required:
    - content
    type: object
  domain.User:
    properties:
      city:
        $ref: '#/definitions/domain.City'
      city_id:
        type: integer
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      iin:
        type: string
      is_active:
        type: boolean
      last_name:
        type: string
      phone:
        type: string
      role:
        $ref: '#/definitions/domain.UserRole'
      role_id:
        type: integer
      updated_at:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API для платформы аренды квартир Renti.kz
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Renti.kz API
  version: "1.0"
paths:
...
  /chat/messages/{messageId}:
    delete:
      description: Удаляет сообщение (только автор может удалить свое сообщение)
      parameters:
      - description: ID сообщения
        in: path
        name: messageId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Сообщение удалено
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Ошибка при удалении
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Удаление сообщения
      tags:
      - chat
    put:
      consumes:
      - application/json
      description: Редактирует содержимое сообщения (только автор может редактировать
        свои сообщения)
      parameters:
      - description: ID сообщения
        in: path
        name: messageId
        required: true
        type: integer
      - description: Новый текст сообщения
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Сообщение обновлено
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Неверные данные запроса
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Редактирование сообщения
      tags:
      - chat
  /chat/rooms:
    get:
      description: Получает список комнат чата пользователя с пагинацией
      parameters:
      - description: Фильтр по статусу комнаты (pending, active, closed)
        in: query
        name: status
        type: string
      - default: 1
        description: Номер страницы
        in: query
        name: page
        type: integer
      - default: 20
        description: Размер страницы
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список комнат чата
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/domain.ChatRoom'
                type: array
              pagination:
                type: object
              success:
                type: boolean
            type: object
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Список комнат чата
      tags:
      - chat
    post:
      consumes:
      - application/json
      description: Создает новую комнату чата для бронирования с консьержем
      parameters:
      - description: Данные для создания комнаты
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateChatRoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Комната чата создана
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.ChatRoom'
              type: object
        "400":
          description: Неверные данные запроса
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Создание комнаты чата
      tags:
      - chat
  /chat/rooms/{roomId}:
    get:
      description: Получает информацию о комнате чата по ID
      parameters:
      - description: ID комнаты чата
        in: path
        name: roomId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Комната чата
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.ChatRoom'
              type: object
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получение комнаты чата
      tags:
      - chat
  /chat/rooms/{roomId}/activate:
    post:
      consumes:
      - application/json
      description: Активирует комнату чата (переводит из pending в active)
      parameters:
      - description: ID комнаты
        in: path
        name: roomId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Активировать чат
      tags:
      - chat
  /chat/rooms/{roomId}/close:
    post:
      consumes:
      - application/json
      description: Закрывает комнату чата (переводит в closed)
      parameters:
      - description: ID комнаты
        in: path
        name: roomId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Закрыть чат
      tags:
      - chat
  /chat/rooms/{roomId}/messages:
    get:
      description: Получает список сообщений из указанной комнаты чата с пагинацией
      parameters:
      - description: ID комнаты чата
        in: path
        name: roomId
        required: true
        type: integer
      - default: 1
        description: Номер страницы
        in: query
        name: page
        type: integer
      - default: 50
        description: Размер страницы
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список сообщений
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/domain.ChatMessage'
                type: array
              pagination:
                type: object
              success:
                type: boolean
            type: object
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получение сообщений
      tags:
      - chat
    post:
      consumes:
      - application/json
      description: Отправляет сообщение в указанную комнату чата
      parameters:
      - description: ID комнаты чата
        in: path
        name: roomId
        required: true
        type: integer
      - description: Данные сообщения
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.SendMessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Сообщение отправлено
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.ChatMessage'
              type: object
        "400":
          description: Неверные данные запроса
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Отправка сообщения
      tags:
      - chat
  /chat/rooms/{roomId}/read:
    post:
      consumes:
      - application/json
      description: Отмечает указанные сообщения в комнате как прочитанные
      parameters:
      - description: ID комнаты чата
        in: path
        name: roomId
        required: true
        type: integer
      - description: Список ID сообщений
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.MarkMessagesReadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Сообщения отмечены как прочитанные
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Неверные данные запроса
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Отметить сообщения как прочитанные
      tags:
      - chat
  /chat/rooms/{roomId}/unread-count:
    get:
      description: Получает количество непрочитанных сообщений в комнате чата
      parameters:
      - description: ID комнаты чата
        in: path
        name: roomId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Количество непрочитанных
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  properties:
                    unread_count:
                      type: integer
                  type: object
              type: object
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получить количество непрочитанных сообщений
      tags:
      - chat
...
  /concierge/chat/rooms:
    get:
      description: Получает список чат-комнат для консьержа
      parameters:
      - description: Фильтр по статусу
        in: query
        name: status
        type: string
      - default: 1
        description: Номер страницы
        in: query
        name: page
        type: integer
      - default: 20
        description: Размер страницы
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Чаты консьержа
          schema:
            properties:
              data:
                properties:
                  rooms:
                    items:
                      $ref: '#/definitions/domain.ChatRoom'
                    type: array
                  total:
                    type: integer
                type: object
              success:
                type: boolean
            type: object
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получение чатов консьержа
      tags:
      - concierge-interface
...
securityDefinitions:
  ApiKeyAuth:
    description: Bearer token для авторизации
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"